Deleted Tower Identiy Class.
Was planning to use a composite key to make sure a user isn't able to
create the same tower twice. However after some research found it was
better to perform a check before inserting into a dictionary using the
backend (hence preventing the exception and rendering the composite key pointless).