Working on #25 Upgraded and added Geo/History building to Unity Project
refined the model and addd it to the unity project. Also modified the lighting to increase the intensity of the global light and change the direction and lowered the intensity of a secondary light. This secondary light adds ambient light to the scene so when somethings in shadow it is still lit by something