Working on #25 and #61 Created Maths and Bio Building
Created a glitchy cube building for maths. Created stage 1 Helix building for Biology/Chemistry. In the same helix building blender file 'DNA Buidling.blend' there is a work in progress stage 2 helix building. In the maths and physics folder there is also a work in progress space shuttle for the physics building. The maths and bio building were added to the Unity Project