Working on #26 Created the all the gradient textures in blender.
Unsure how to create gradient materials in Unity. Will have to look up how to make shaders. In blender shaders are very sttraightforward and would make it super easy to apply images and maps later on so I created the gradient materials and will look up ways of exporting these materials to Unity. It may not pan out but worth a try