Working on #25 and #61 Optimised the glitch cube maths building and reshaped the Biology building
The maths building was starting to cause some lag as there were alot of hidden cubes and hidden faces due to me using a partical system to randomly distribute the blocks. this has no been resolved. The biology building had too many floors for its height relative to other buildings so each floor is not taller. the building is the same height. the biology building now also has a base with an entrance.