Working on #26 Installed Unity Lightweight Render Pipeline. Upgraded Materials and experimented with Gradient and Procedural Water
Okay so this is a long one... We want maximum customisation in this game. To do so, the most efficient way of creating materials is using a graph node editor, like in blender, you can preview the material whilst you work on it and there are lots of inbuilt functions. To use materials with custom shaders (graph shaders) you need to install the Light weight render pipeline (best for android). As a result, all the standard materials now no longer work. All the materials were upgraded to LWRP materials and everything appears fine. I don't believe this will have to be installed manually on other machines, it should be automatic, if not we will have a meeting to discuss and I will create a wiki page about it. During this process I also noticed that there were examples in TextMesh Pro and removed them as it took up a fair ammount of space for the project.