Scaled up the image texture for the terrain shading
The path was a seperate material and contrasted with the baked in shadows of the terrain quite alot. I made the pathway use the same image texture as the regular terrain, however it was quite low res. I upscaled the baked lighting and shadows texture to 8k. This may be too high for mobile, but can be reduced in Unity to 4k if it runs poorly