Working on #25 and #26 Combined all objects in each model into a single object per model for material customisation reasons
The material customisation changed the 1st and second materials of each object in a model to the primary and secondary materials. If an object does not have the primary and secondary materials as the 1st and second material (e.g the object only has the secondary material then this material will be assigned the primary material, regardless of the intention. Now all the objects are combined together, the single model object will have every material in the expected order, and this will improve efficiency when changing a models materials as it will only have to iterate through one object instead of 10s of objects.