Working on #75 - Materials
Added the Load_Customisation script to all of the parent objects in the main scene. Updated the Color_Picker UI script to store the users choices in a temp variable, and created a confirm button script that applies the temp variables to the persistant list and returns to the main scene.