Working on #74. Modified application.properties to fix connection issues
After changing the key constraints the database was experiencing bugs.
In order to recitify this I modified the source file to use a fresh
database. Also to allow smoother testing as well as to avoid overlapping
issues, the databases behaviour is now changed to "create-drop". This
means it will recreate the schema every time it's started and destroy it
(along with all the data) once its closed.
Obviously this is purely for testing purposes.